refactor: update imports to use shadcn directory in task-related pages
